requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pytest>=6.2.5
matplotlib>=3.4.0
plotly>=5.3.0
bokeh>=2.4.0
requests>=2.26.0
sqlalchemy>=1.4.0
pyyaml>=5.4.0
python-dateutil>=2.8.2
scikit-learn>=1.0.0
xarray>=2022.3.0
seaborn>=0.11.0 
holoviews>=1.15.0
prophet>=1.0.0
pyvis>=0.3.0
pymongo>=4.0.0
jsonschema>=4.0.0